+++ src/tests/test_definitions.py
@@ -19,6 +19,7 @@ from pydocstyle.checker import check
     'superfluous_quotes',
     'noqa',
     'sections',
+    'functions',
     'canonical_google_examples',
     'canonical_numpy_examples',
     'canonical_pep257_examples',