+++ docs/release_notes.rst
@@ -33,6 +33,8 @@ Bug Fixes
 * Fix decorator parsing (#411).
 * Google-style sections no longer cause false errors when used with
   Numpy-style sections (#388, #424).
+* D202: Allow a blank line after function docstring when followed by
+  declaration of an inner function or class (#395, #426).
 * Fix D401 and D404 checks not working for docstrings containing only one word and ending with non-alpha character (#421)
 
 4.0.1 - August 14th, 2019
@@ -47,7 +49,6 @@ Bug Fixes
 * Fixed RST error in table which resulted in the online documentation missing
   the violation code table (#396).
 * Fixed IndentationError when parsing function arguments (#392).
-
 
 4.0.0 - July 6th, 2019
 ----------------------